still working on shaders

+++ LumEngine/engine/include/modules/rhi/core/rhi_texture.hpp
@@ -11,28 +11,31 @@ namespace lum::rhi {
 
 	};
 
+	// Filter that applies to texture, when texture is magnified ( scaled up )
 	enum class TextureMagFilter : uint8_t {
 		Linear,	// Takes 4 neighbouring texels and calculates it weighted average value (Effect: Smoothness, Fuzzy, no sharp edges).
 		Nearest	// Takes texture coordinate from UV map, rounds it and takes 1 texel (Effect: Sharpness). 
 	};
 
+	// Filter that applies to texture, when texture is minified ( scaled down )
 	enum class TextureMinFilter : uint8_t {
-		Linear,					// Takes 4 neighbouring texels and calculates it weighted average value (Effect: Smoothness, Fuzzy, no sharp edges).
-		Linear_mipmap_nearest,	// Takes 4 neighbouring texels from single mipmap and calculates it weighted average value (Effect: Old style games, smooth).
-		Linear_mipmap_linear,	// Takes 4 neighbouring texels from 2 mipmaps and calculates it weighted average value (Effect: Best quality, smooth but not fuzzy).
-		Nearest,				// Takes texture coordinate from UV map, rounds it and takes 1 texel (Effect: Sharpness). 
-		Nearest_mipmap_nearest, // Takes 1 texel from single mipmap (Effect: Even more sharpness, worse quality).
-		Nearest_mipmap_linear	// Takes 1 texel from 2 mipmaps and calculates it weighet average value (Effect: Low detailed, weird look).
+		Linear,					// Takes 4 neighbouring texels and calculates it weighted average value ( Effect: Smoothness, Fuzzy, no sharp edges, NO MIPMAP ).
+		Linear_mipmap_nearest,	// Takes 4 neighbouring texels from single mipmap and calculates it weighted average value ( Effect: Compromise for quality and performance, smoothness ).
+		Linear_mipmap_linear,	// Takes 4 neighbouring texels from 2 mipmaps and calculates it weighted average value ( Effect: Best quality, smooth but not fuzzy ).
+		Nearest,				// Takes texture coordinate from UV map, rounds it and takes 1 texel ( Effect: Sharpness, NO MIPMAP ). 
+		Nearest_mipmap_nearest, // Takes 1 texel from single mipmap ( Effect: Even more sharpness, worse quality ).
+		Nearest_mipmap_linear	// Takes 1 texel from 2 mipmaps and calculates it weighet average value ( Effect: Low detailed, weird look ).
 	};
 
 	struct TextureDescriptor {
 
-		// Filter that applies to texture, when texture is magnified (scaled up)
+		// Filter that applies to texture, when texture is magnified ( scaled up )
 		TextureMagFilter mag_filter{};
-		// Filter that applies to texture, when texture is minified (scaled down)
+		// Filter that applies to texture, when texture is minified ( scaled down )
 		TextureMinFilter min_filter{};
 
-		float anisotropy = 0;
+		// Level of anisotropic filtering ( sharpens textures at steep viewing angles, prevents blur and stretch )
+		int anisotropy = 0;
 
 		// Path/Name of the texture
 		cstr filename = nullptr;
